Preserve correct __LINE__ iin Kernel#binding
This allows introspection tools such as Pry determine accurately the
__LINE__ and __FILE__ on which the snapshot of the binding was taken.

Before this change we were using the value of binding.getLine() as a
substitute for the line on which Kernel#binding was called, however this gave
the wrong results. Now we hook into the same mechanism as Kernel#caller to
ensure the line number is correct.

Specs for this behaviour were added at
https://github.com/jruby/rubyspec/commit/88f3dff4.